Please Read this.
# Assignment-No3-Os
Sir Please Read this, As this is about the assignment which was supposed to be submitted on firday Oct,19 2018.
I was late in the class on Friday because I was making the written Assignment. When I arrived in the Class I was told
that submission of the Assignment would be on Monday. So I thought that the submission of the Codinb part would also be 
on Monday that's why I didn't submitted that Assignment on Friday. I'm so Sorry sir. I wasn't been late but It was only
by a mistake.